third_party/WebKit/LayoutTests/fast/scrolling/fractional-scroll-height-chaining-expected.txt

Verifies that main-thread scrolls chain correctly from a fractional-height scroller to its parent.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS scrollY is 40
PASS successfullyParsed is true

TEST COMPLETE
